+++ apps/frontend/src/hooks/feed/useFeedInfiniteQuery.ts
@@ -3,7 +3,7 @@
 import { useInfiniteQuery } from '@tanstack/react-query';
 import { getFeedPosts, type GetFeedPostsResponse } from '@/apis/feed/feed';
 
-export const useFeedInfiniteQuery = (limit = 10) => {
+export const useFeedInfiniteQuery = (limit = 10, initialData?: GetFeedPostsResponse | null) => {
   return useInfiniteQuery<
     GetFeedPostsResponse,
     Error,
@@ -14,6 +14,12 @@ export const useFeedInfiniteQuery = (limit = 10) => {
     queryKey: ['feed', { limit }],
     queryFn: ({ pageParam }) => getFeedPosts({ cursor: pageParam, limit }),
     initialPageParam: null,
+    initialData: initialData
+      ? {
+          pages: [initialData],
+          pageParams: [null],
+        }
+      : undefined,
     getNextPageParam: (lastPage) => lastPage.nextCursor,
   });
 };
